Check sync status to find missing commits for push
Fixes bug where killing and restarting a push would complete
immediately, which was caused by the check for missing commits testing
for the existence of the commit nodes in the server's tree rather than whether
these nodes were synced. Oxen push now makes a request to the server to
mark each each commit as synced after all the files are uploaded, and
tests for this instead to determine missing commits. This means that the
commits will remain 'missing' until all files have been pushed, allowing
for pushes to restart