Implement dir-level sync for push
Oxen push now marks nodes as synced while uploading file contents to
version store in push_entries. The uploads for large and small files
are now consolidated into one loop. Prior to the upload, each node's
parents and children are collected and tracked to determine when to mark
each node as synced. The hashes of nodes with all their children
uploaded are included with outgoing payloads of file contents