Fix push and workspace_add benchmarks
Fixed issue where the push and workspace_add benchmarks would repeatedly
push/add to the same remote and the same workspace respectively. New
remotes are now created and configured for the push benchmark on every
iteration, and new workspaces are created for the workspace_add benchmark.

This does introduce an issue whereby the push benchmark pushes to an
emprt repository, as opposed to one already filled with files like the
add and workspace_add benchmarks do, but without a clean way to undo a
push, this seems necessary to avoid blowing up setup costs for the
benchmark