Refactor workspace add to use MPSC pipeline
Rebuilds the core logic for adding small files to workspaces,
implementing an MPSC pipeline to maximize utilization of the client.
The producer builds multiparts, offloading the work to a
blocking thread pool, while the consumer uploads the multiparts and
then stages the nodes. Err files are collected from both, which are
logged.