Implement new endpoint for one file workspace add
Implements new endpoint for workspace add, which upload_single_file now
calls. This endpoint uploads the file contents to version store and
stages the nodes in one call, skipping writing the file contents to
the workspace itself to increase efficiency.

Also, implements method to directly add bytes to a PyWorkspace, making
it possible to skip writing buffer contents to disk when uploading them
to a workspace