Implement backpressure and retry for workspace add
Uses a semaphore to limit the amount of concurrent uploads allowed in
the core logic of parallel_batched_small_file_upload. The amount
defaults to 30, but can be configured via an environment variable.

Refactors workspace add to collect files that fail at any point of the
workspace add pipeline and retry them at the end of the function.
Currently, a single file failing after this retry will result in the
operation returning an error