Implement glob_opts for glob path parsing
Introduces the GlobOpts struct, which allows functions to control where
glob paths are matched against. Also includes corrections for unit
tests, including having non-glob dirs be walked for paths if walk_dirs
is set, which is necessary for workspace add. As well, corrections for
different current working directories are included