Annotate controller modules with utoipa
Annotates the functions in workspace, data frame, repository, namespace,
commit, metadata, and file controller modules. This allows us to
automatically generate http docs for the endpoints associated with these
endpoints. These docs contain the request and response structures, as
well as auto-generated code to use them in a variety of languages.

Using utoipa for the docs requires implementing utoipa's traits for
effectively all of the structs that these modules use, requiring changes
to much of the model and view modules