Implement retry logic for version file upload
Implements a retry for the version file upload in workspace add. This
sends the paths of the batch to the old version file upload logic in
api::client::versions, which computes the multiparts and uploads all in
one. If individual err_files are returned, only those file parts are rebuiltand sent. If the operation as a whole failed, all file parts are resent

Also implements configurations for the number of retries allowed and the
http request timeout via environment variables