Implement endpoint to download dir contents as zip
Implements new client and server modules to support downloading dir
contents as a zip file, rather than just as compressed tarballs. This
requires the previous batch_download logic to be split into two separate
functions, one for tarballs and the other for zip files, and the use of
the async_zip library.